HomeWork3/requirements.txt
kafka-python>=2.0.2
faker>=24.0.0
pymongo>=4.6.0
apache-flink>=1.18.0
jdk4py==17.0.9.2
